+++ test/jest/formElement.test.js
@@ -547,13 +547,13 @@ describe("formElement", () => {
       el.$isPending = true;
       await h.wait();
       expect(el.outerHTML).toMatchInlineSnapshot(
-        `"<wup-form role="form" disabled=""><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off" disabled=""><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off" disabled=""><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off" disabled=""><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" disabled="" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
+        `"<wup-form role="form" aria-busy="true" readonly=""><wup-text busy=""><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off" readonly=""><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off" readonly=""><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off" readonly=""><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
       );
 
       el.changePending(true); // just for coverage
       await h.wait();
       expect(el.outerHTML).toMatchInlineSnapshot(
-        `"<wup-form role="form" disabled=""><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off" disabled=""><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off" disabled=""><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off" disabled=""><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" disabled="" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
+        `"<wup-form role="form" aria-busy="true" readonly=""><wup-text busy=""><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off" readonly=""><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off" readonly=""><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off" readonly=""><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
       );
 
       el.$isPending = false;
@@ -563,16 +563,15 @@ describe("formElement", () => {
       );
       el.$isPending = false; // just for coverage
 
-      const submitFn = (e) => (e.detail.waitFor = new Promise((resolve) => setTimeout(() => resolve("true"), 500)));
+      const submitFn = (e) => (e.detail.waitFor = new Promise((res) => setTimeout(() => res("true"), 500)));
       el.addEventListener("$submit", submitFn);
       btnSubmit.click();
       await h.wait(1);
       expect(el.$isPending).toBe(true);
       expect(el.outerHTML).toMatchInlineSnapshot(
-        `"<wup-form role="form"><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off"><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off"><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off"><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" disabled=""></button><wup-spin style="display: none;"><div></div></wup-spin></wup-form>"`
+        `"<wup-form role="form" aria-busy="true"><wup-text busy=""><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off"><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off"><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off"><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit"></button><wup-spin style="display: none;"><div></div></wup-spin></wup-form>"`
       );
       await h.wait();
-      await h.wait(1);
       expect(el.$isPending).toBe(false);
       expect(el.outerHTML).toMatchInlineSnapshot(
         `"<wup-form role="form"><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off"><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off"><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off"><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit"></button></wup-form>"`
@@ -580,12 +579,11 @@ describe("formElement", () => {
       el.removeEventListener("$submit", submitFn);
 
       // check again with $onSubmit
-      el.$onSubmit = () => new Promise((resolve) => setTimeout(() => resolve("true"), 500));
+      el.$onSubmit = () => new Promise((res) => setTimeout(() => res("true"), 500));
       btnSubmit.click();
       await h.wait(1);
       expect(el.$isPending).toBe(true);
       await h.wait();
-      await h.wait(1);
       expect(el.$isPending).toBe(false);
     });
 
@@ -730,16 +728,43 @@ describe("formElement", () => {
       });
 
       test("lockOnPending", async () => {
-        el.$onSubmit = () => new Promise((resolve) => setTimeout(resolve, 300));
+        const onSubmit = jest.fn().mockImplementation(() => new Promise((res) => setTimeout(res, 300)));
+        el.$onSubmit = onSubmit;
         el.$options.submitActions = SubmitActions.lockOnPending;
+        btnSubmit.focus();
         btnSubmit.click();
         await h.wait(50);
         expect(el.$isPending).toBe(true);
-        expect(el.$options.disabled).toBe(true);
+        expect(el.$options.disabled).toBe(false); // otherwise focus can be missed
+        expect(document.activeElement).toBe(btnSubmit);
         expect(el.outerHTML).toMatchInlineSnapshot(
-          `"<wup-form role="form" disabled=""><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off" disabled=""><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off" disabled=""><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off" disabled=""><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" disabled="" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
+          `"<wup-form role="form" aria-busy="true" readonly=""><wup-text busy=""><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off" readonly=""><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off" readonly=""><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text busy=""><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off" readonly=""><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
         );
         await h.wait();
+        expect(el.$isPending).toBe(false);
+        expect(onSubmit).toBeCalledTimes(1);
+
+        // again + 2 clicks
+        jest.clearAllMocks();
+        btnSubmit.click();
+        await h.wait(50);
+        btnSubmit.click(); // 2nd time must be prevented
+        await h.wait(50);
+        expect(el.$isPending).toBe(true);
+        expect(onSubmit).toBeCalledTimes(1);
+        await h.wait();
+        expect(el.$isPending).toBe(false);
+
+        // again + 2 Enter-s
+        jest.clearAllMocks();
+        await h.userPressKey(el, { key: "Enter" });
+        await h.wait(50);
+        await h.userPressKey(el, { key: "Enter" });
+        await h.wait(50);
+        expect(el.$isPending).toBe(true);
+        expect(onSubmit).toBeCalledTimes(1);
+        await h.wait();
+        expect(el.$isPending).toBe(false);
 
         el.$options.submitActions = SubmitActions.none;
         btnSubmit.click();
@@ -747,7 +772,7 @@ describe("formElement", () => {
         expect(el.$isPending).toBe(true);
         expect(el.$options.disabled).toBeFalsy();
         expect(el.outerHTML).toMatchInlineSnapshot(
-          `"<wup-form role="form"><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off"><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off"><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off"><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" disabled="" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
+          `"<wup-form role="form"><wup-text><label for="txt1"><span><input placeholder=" " type="text" id="txt1" autocomplete="off"><strong>Email</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt2"><span><input placeholder=" " type="text" id="txt2" autocomplete="off"><strong>First Name</strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><wup-text><label for="txt3"><span><input placeholder=" " type="text" id="txt3" autocomplete="off"><strong></strong></span><button wup-icon="" clear="" tabindex="-1" aria-hidden="true" type="button"></button></label></wup-text><button type="submit" aria-busy="true"></button><wup-spin style="position: absolute; display: none;" aria-label="Loading. Please wait"><div></div></wup-spin></wup-form>"`
         );
         await h.wait();
       });